seperate format tests from value tests
